Disable Watchdog
The Watchdog was an important part of FRC coe, but currently hard to implement for FTC. Look into doing this in the future.

Within the FTCRobotControllerActivity (found in FtcRootController.src.main.java.internal), there is a controllerService class that contains a robot and event manager. The event manger contains information about the robots "heart beat". This might be beneficial. As of 9/2/2020, it's illegal to modify Robot, event manager, or controllService, but the objects within FtcRobotController Activity can be made public and then used by other classes.