Remove Robotstate
This should be added back in to eventually so that we can make sure commands are being cancelled on disabled. Could not figure out how to get the FtcContollerActivity object, or where it was used. If that could be determined, it could either be made a sigleton with a get instance class or the class could be pulled to get the controllerService.